Add methods for grain diff and grain estimation
Grain diff compares an original source and a denoised source and
generates a film grain table based on the difference. Grain estimation
analyzes a plane from an input file (can be used for luma and/or chroma)
and generates an estimate of the grain strength as a float value.

Grain diff is more accurate, whereas grain estimation should be
considerably faster.